% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{asap_bbknn}
\alias{asap_bbknn}
\title{Reconcile multi-batch matrices by batch-balancing KNN}
\usage{
asap_bbknn(
  data_nk_vec,
  row_names_vec,
  KNN_PER_BATCH = 3L,
  BLOCK_SIZE = 1000L,
  NUM_THREADS = 1L,
  verbose = TRUE
)
}
\arguments{
\item{data_nk_vec}{a list of sample x factor matrices}

\item{row_names_vec}{a list of sample x 1 names}

\item{KNN_PER_BATCH}{(default: 3)}

\item{BLOCK_SIZE}{each parallel job size (default: 100)}

\item{NUM_THREADS}{number of parallel threads (default: 1)}

\item{verbose}{(default: TRUE)}
}
\value{
a list that contains:
\itemize{
\item adjusted (N x K) matrix
\item bbknn batch-balanced kNN adjacency matrix
\item batches batch membership
\item knn edges
}
}
\description{
Reconcile multi-batch matrices by batch-balancing KNN
}
