% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{asap_build_interacting_columns}
\alias{asap_build_interacting_columns}
\title{Identify pairs of columns interacting with one another}
\usage{
asap_build_interacting_columns(
  y_dn,
  z_dm,
  log_beta,
  beta_row_names,
  knn = 10L,
  do_stdize_beta = TRUE,
  do_log1p = FALSE,
  verbose = TRUE,
  NUM_THREADS = 1L,
  CELL_NORM = 10000,
  BLOCK_SIZE = 1000L
)
}
\arguments{
\item{y_dn}{sparse data matrix (D x N)}

\item{z_dm}{sparse data matrix (D x M)}

\item{log_beta}{D x K log dictionary/design matrix}

\item{beta_row_names}{row names log_beta (D vector)}

\item{knn}{How many nearest neighbours we want (default: 10)}

\item{do_stdize_beta}{use standardized log_beta (Default: TRUE)}

\item{do_log1p}{do log(1+y) transformation}

\item{verbose}{verbosity}

\item{NUM_THREADS}{number of threads in data reading}

\item{CELL_NORM}{sample normalization constant (default: 1e4)}

\item{BLOCK_SIZE}{disk I/O block size (number of columns)}
}
\description{
Identify pairs of columns interacting with one another
}
