% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{asap_build_interaction}
\alias{asap_build_interaction}
\title{Identify pairs of columns interacting with one another}
\usage{
asap_build_interaction(
  y_dn,
  z_dm,
  log_beta,
  beta_row_names,
  knn = 10L,
  r_log_delta = NULL,
  do_stdize_beta = TRUE,
  do_log1p = FALSE,
  verbose = TRUE,
  NUM_THREADS = 1L,
  BLOCK_SIZE = 100L
)
}
\arguments{
\item{y_dn}{sparse data matrix (D x N)}

\item{z_dm}{sparse data matrix (D x M)}

\item{log_beta}{D x K log dictionary/design matrix}

\item{beta_row_names}{row names log_beta (D vector)}

\item{knn}{How many nearest neighbours we want (default: 10)}

\item{r_log_delta}{D x B log batch effect matrix}

\item{do_stdize_beta}{use standardized log_beta (Default: TRUE)}

\item{do_log1p}{do log(1+y) transformation}

\item{verbose}{verbosity}

\item{NUM_THREADS}{number of threads in data reading}

\item{BLOCK_SIZE}{disk I/O block size (number of columns)}
}
\description{
Identify pairs of columns interacting with one another
}
