% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{asap_build_interaction_mtx}
\alias{asap_build_interaction_mtx}
\title{Identify pairs of columns interacting with one another}
\usage{
asap_build_interaction_mtx(
  mtx_file,
  row_file,
  col_file,
  idx_file,
  log_beta,
  beta_row_names,
  knn = 10L,
  r_log_delta = NULL,
  mtx_file_rhs = NULL,
  row_file_rhs = NULL,
  col_file_rhs = NULL,
  idx_file_rhs = NULL,
  do_stdize_beta = TRUE,
  do_log1p = FALSE,
  verbose = TRUE,
  NUM_THREADS = 1L,
  BLOCK_SIZE = 100L,
  MAX_ROW_WORD = 2L,
  ROW_WORD_SEP = "_",
  MAX_COL_WORD = 100L,
  COL_WORD_SEP = "@"
)
}
\arguments{
\item{mtx_file}{matrix-market-formatted data file (D x N, bgzip)}

\item{row_file}{row names file (D x 1)}

\item{col_file}{column names file (N x 1)}

\item{idx_file}{matrix-market colum index file}

\item{log_beta}{D x K log dictionary/design matrix}

\item{beta_row_names}{row names log_beta (D vector)}

\item{knn}{How many nearest neighbours we want (default: 10)}

\item{r_log_delta}{D x B log batch effect matrix}

\item{mtx_file_rhs}{right-hand-side matrix-market-formatted data file (bgzip)}

\item{row_file_rhs}{right-hand-side row names (gene/feature names)}

\item{col_file_rhs}{right-hand-side column names (cell/column names)}

\item{idx_file_rhs}{right-hand-side matrix-market colum index file}

\item{do_stdize_beta}{use standardized log_beta (Default: TRUE)}

\item{do_log1p}{do log(1+y) transformation}

\item{verbose}{verbosity}

\item{NUM_THREADS}{number of threads in data reading}

\item{BLOCK_SIZE}{disk I/O block size (number of columns)}

\item{MAX_ROW_WORD}{maximum words per line in \code{row_files[i]}}

\item{ROW_WORD_SEP}{word separation character to replace white space}

\item{MAX_COL_WORD}{maximum words per line in \code{col_files[i]}}

\item{COL_WORD_SEP}{word separation character to replace white space}
}
\description{
Identify pairs of columns interacting with one another
}
