% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{asap_deconv_pmf_stat}
\alias{asap_deconv_pmf_stat}
\title{Estimated deconvolved topic statistics}
\usage{
asap_deconv_pmf_stat(
  bulk_dm,
  bulk_row_names,
  Ref_nk,
  log_beta,
  beta_row_names,
  do_stdize_beta = TRUE,
  do_log1p = FALSE,
  verbose = TRUE,
  a0 = 1,
  b0 = 1,
  NUM_THREADS = 1L,
  KNN_PER_SAMPLE = 15L
)
}
\arguments{
\item{bulk_dm}{Convoluted bulk data}

\item{bulk_row_names}{row names of bulk_data (D vector)}

\item{Ref_nk}{reference correlation data (N x K)}

\item{log_beta}{log dictionary matrix (D x K)}

\item{beta_row_names}{row names log_beta (D vector)}

\item{do_stdize_beta}{use standardized log_beta (default: TRUE)}

\item{do_log1p}{do log(1+y) transformation}

\item{verbose}{verbosity}

\item{NUM_THREADS}{number of threads in data reading}

\item{KNN_PER_SAMPLE}{k Nearest Neighbours per bulk sample}
}
\description{
Estimated deconvolved topic statistics
}
