% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{asap_deconv_stat}
\alias{asap_deconv_stat}
\title{Topic deconvolution}
\usage{
asap_deconv_stat(
  bulk_dm,
  bulk_row_names,
  log_beta,
  beta_row_names,
  Ref_nk,
  rseed = 42L,
  verbose = TRUE,
  NUM_THREADS = 1L,
  KNN_PER_SAMPLE = 15L,
  BATCH_ADJ_ITER = 100L,
  a0 = 1,
  b0 = 1
)
}
\arguments{
\item{bulk_dm}{Convoluted bulk data}

\item{bulk_row_names}{row names of bulk_data (D vector)}

\item{log_beta}{log dictionary matrix (D x K)}

\item{beta_row_names}{row names log_beta (D vector)}

\item{Ref_nk}{reference correlation data (N x K)}

\item{do_stdize_beta}{use standardized log_beta (default: TRUE)}
}
\description{
Topic deconvolution
}
