% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{asap_fit_modular_nmf}
\alias{asap_fit_modular_nmf}
\title{Non-negative matrix factorization with the row modules}
\usage{
asap_fit_modular_nmf(
  Y,
  maxK,
  maxL,
  collapsing = NULL,
  mcem = 100L,
  burnin = 10L,
  latent_iter = 1L,
  degree_iter = 1L,
  thining = 3L,
  verbose = TRUE,
  eval_llik = TRUE,
  a0 = 1,
  b0 = 1,
  rseed = 42L,
  NUM_THREADS = 1L,
  update_loading = TRUE,
  gibbs_sampling = TRUE
)
}
\arguments{
\item{Y}{data matrix (gene x sample)}

\item{maxK}{maximum number of factors}

\item{maxL}{maximum number of row modules}

\item{collapsing}{L x row collapsing matrix (L < gene)}

\item{mcem}{number of Monte Carl Expectation Maximization}

\item{burnin}{burn-in period}

\item{latent_iter}{latent sampling steps}

\item{degree_iter}{row and column degree optimization steps}

\item{thining}{thining interval in record keeping}

\item{verbose}{verbosity}

\item{eval_llik}{evaluate log-likelihood}

\item{a0}{gamma(a0, b0)}

\item{b0}{gamma(a0, b0)}

\item{rseed}{random seed}

\item{NUM_THREADS}{number of parallel jobs}
}
\description{
Non-negative matrix factorization with the row modules
}
