% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{asap_fit_nmf_alternate}
\alias{asap_fit_nmf_alternate}
\title{A quick NMF estimation based on alternating Poisson regressions}
\usage{
asap_fit_nmf_alternate(
  Y_dn,
  maxK,
  max_iter = 100L,
  burnin = 10L,
  verbose = TRUE,
  a0 = 1,
  b0 = 1,
  rseed = 42L,
  EPS = 1e-04
)
}
\arguments{
\item{Y_dn}{non-negative data matrix (gene x sample)}

\item{maxK}{maximum number of factors}

\item{max_iter}{number of variation Expectation Maximization steps}

\item{burnin}{number of optimization steps w/o scaling}

\item{verbose}{verbosity}

\item{a0}{gamma(a0, b0)}

\item{b0}{gamma(a0, b0)}

\item{rseed}{random seed}
}
\value{
a list that contains:
\itemize{
\item log.likelihood log-likelihood trace
\item beta dictionary (gene x factor)
\item log.beta log-dictionary (gene x factor)
\item theta loading (sample x factor)
\item log.theta log-loading (sample x factor)
\item log.phi auxiliary variables (gene x factor)
\item log.rho auxiliary variables (sample x factor)
}
}
\description{
A quick NMF estimation based on alternating Poisson regressions
}
