% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{asap_fit_pmf}
\alias{asap_fit_pmf}
\title{A quick PMF estimation based on alternating Poisson regressions}
\usage{
asap_fit_pmf(
  Y_,
  maxK,
  max_iter = 100L,
  r_A_dd_list = NULL,
  r_A_nn_list = NULL,
  burnin = 0L,
  verbose = TRUE,
  a0 = 1,
  b0 = 1,
  do_log1p = FALSE,
  rseed = 1337L,
  svd_init = FALSE,
  EPS = 1e-08,
  NUM_THREADS = 0L
)
}
\arguments{
\item{Y_}{non-negative data matrix (gene x sample)}

\item{maxK}{maximum number of factors}

\item{max_iter}{max number of optimization steps}

\item{burnin}{number of initiation steps (default: 50)}

\item{verbose}{verbosity}

\item{a0}{gamma(a0, b0) default: a0 = 1}

\item{b0}{gamma(a0, b0) default: b0 = 1}

\item{do_log1p}{do log(1+y) transformation}

\item{rseed}{random seed (default: 1337)}

\item{svd_init}{initialize by SVD (default: FALSE)}

\item{EPS}{(default: 1e-8)}

\item{min_iter}{min number of optimization steps}

\item{do_scale}{scale each column by standard deviation (default: TRUE)}
}
\value{
a list that contains:
\itemize{
\item log.likelihood log-likelihood trace
\item theta loading (sample x factor)
\item log.theta log-loading (sample x factor)
\item log.theta.sd sd(log-loading) (sample x factor)
\item beta dictionary (gene x factor)
\item log.beta log dictionary (gene x factor)
\item log.beta.sd sd(log-dictionary) (gene x factor)
}
}
\description{
A quick PMF estimation based on alternating Poisson regressions
}
