% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{asap_fit_pmf_cbind}
\alias{asap_fit_pmf_cbind}
\title{A quick PMF estimation based on alternating Poisson regressions
while sharing a dictionary/factors matrix}
\usage{
asap_fit_pmf_cbind(
  y_dn_vec,
  maxK,
  max_iter = 100L,
  verbose = TRUE,
  a0 = 1,
  b0 = 1,
  do_log1p = FALSE,
  rseed = 1337L,
  EPS = 1e-08,
  NUM_THREADS = 0L
)
}
\arguments{
\item{y_dn_vec}{a list of non-negative data matrices (gene x sample)}

\item{maxK}{maximum number of factors}

\item{max_iter}{max number of optimization steps}

\item{verbose}{verbosity}

\item{a0}{gamma(a0, b0) default: a0 = 1}

\item{b0}{gamma(a0, b0) default: b0 = 1}

\item{do_log1p}{do log(1+y) transformation}

\item{rseed}{random seed (default: 1337)}

\item{EPS}{(default: 1e-8)}

\item{min_iter}{min number of optimization steps}

\item{do_scale}{scale each column by standard deviation (default: TRUE)}
}
\value{
a list that contains:
\itemize{
\item log.likelihood log-likelihood trace
\item beta dictionary (gene x factor)
\item log.beta log-dictionary (gene x factor)
\item log.beta.sd sd(log-dictionary) (gene x factor)
\item theta a list of loading matrices (sample x factor)
\item log.theta a list of log loadings (sample x factor)
\item log.theta.sd a list of standard deviations (sample x factor)
}
}
\description{
A quick PMF estimation based on alternating Poisson regressions
while sharing a dictionary/factors matrix
}
