% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{asap_fit_pmf_delta}
\alias{asap_fit_pmf_delta}
\title{A quick PMF estimation based on alternating Poisson regressions
across multiple matrices with the same dimensionality}
\usage{
asap_fit_pmf_delta(
  y_ref,
  y_dn_vec,
  maxK,
  max_iter = 100L,
  verbose = TRUE,
  a0 = 1,
  b0 = 1,
  do_log1p = FALSE,
  rseed = 1337L,
  svd_init = FALSE,
  do_degree_correction = FALSE,
  normalize_cols = FALSE,
  EPS = 1e-08,
  jitter = 1,
  NUM_THREADS = 0L
)
}
\arguments{
\item{y_ref}{reference data matrix (gene x sample)}

\item{y_dn_vec}{a list of non-negative data matrices (gene x sample)}

\item{maxK}{maximum number of factors}

\item{max_iter}{max number of optimization steps}

\item{verbose}{verbosity}

\item{a0}{gamma(a0, b0) default: a0 = 1}

\item{b0}{gamma(a0, b0) default: b0 = 1}

\item{do_log1p}{do log(1+y) transformation}

\item{rseed}{random seed (default: 1337)}

\item{svd_init}{initialize by SVD (default: FALSE)}

\item{normalize_cols}{normalize columns by col_norm (default: FALSE)}

\item{EPS}{(default: 1e-8)}

\item{jitter}{(default: 1)}

\item{min_iter}{min number of optimization steps}
}
\value{
a list that contains:
\itemize{
\item log.likelihood log-likelihood trace
\item theta loading (sample x factor)
\item log.theta log-loading (sample x factor)
\item log.theta.sd sd(log-loading) (sample x factor)
\item beta dictionary (gene x factor)
\item log.beta log dictionary (gene x factor)
\item log.beta.sd sd(log-dictionary) (gene x factor)
\item delta a list of dictionary matrices (gene x factor)
\item log.delta a list of log dictionary (gene x factor)
\item log.delta.sd a list of standard deviations (gene x factor)
}
}
\description{
A quick PMF estimation based on alternating Poisson regressions
across multiple matrices with the same dimensionality
}
