% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{asap_fit_pmf_linking}
\alias{asap_fit_pmf_linking}
\title{Estimate two-layered PMF (experimental)}
\usage{
asap_fit_pmf_linking(
  X_,
  Y_,
  maxK,
  max_iter = 100L,
  burnin = 0L,
  verbose = TRUE,
  a0 = 1,
  b0 = 1,
  do_log1p = FALSE,
  rseed = 1337L,
  svd_init = FALSE,
  EPS = 1e-08,
  NUM_THREADS = 0L
)
}
\description{
Estimate two-layered PMF (experimental)
}
