% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{asap_interaction_pmf_stat_mtx}
\alias{asap_interaction_pmf_stat_mtx}
\title{Topic statistics to estimate factor loading}
\usage{
asap_interaction_pmf_stat_mtx(
  mtx_file,
  row_file,
  col_file,
  idx_file,
  log_beta,
  beta_row_names,
  W_nm_list,
  mtx_file2 = NULL,
  row_file2 = NULL,
  col_file2 = NULL,
  idx_file2 = NULL,
  A_dd_list = NULL,
  do_stdize_beta = TRUE,
  do_product = FALSE,
  NUM_THREADS = 0L,
  BLOCK_SIZE = 1000L,
  MAX_ROW_WORD = 2L,
  ROW_WORD_SEP = "_",
  MAX_COL_WORD = 100L,
  COL_WORD_SEP = "@",
  verbose = FALSE
)
}
\arguments{
\item{mtx_file}{matrix-market-formatted data file (D x N, bgzip)}

\item{row_file}{row names file (D x 1)}

\item{col_file}{column names file (N x 1)}

\item{idx_file}{matrix-market colum index file}

\item{log_beta}{D x K log dictionary/design matrix}

\item{beta_row_names}{row names log_beta (D vector)}

\item{W_nm_list}{list(src.index, tgt.index, \link{weights}) for columns}

\item{A_dd_list}{list(src.index, tgt.index, \link{weights}) for features}

\item{do_stdize_beta}{use standardized log_beta (default: TRUE)}

\item{do_product}{yi * yj for interaction (default: FALSE)}

\item{NUM_THREADS}{number of threads in data reading}

\item{BLOCK_SIZE}{disk I/O block size (number of columns)}

\item{MAX_ROW_WORD}{maximum words per line in \code{row_files[i]}}

\item{ROW_WORD_SEP}{word separation character to replace white space}

\item{MAX_COL_WORD}{maximum words per line in \code{col_files[i]}}

\item{COL_WORD_SEP}{word separation character to replace white space}

\item{verbose}{verbosity}
}
\value{
a list that contains:
\itemize{
\item beta dictionary matrix (row x factor)
\item corr empirical correlation (column x factor)
\item colsum the sum of each column (column x 1)
}
}
\description{
Topic statistics to estimate factor loading
}
