% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{asap_nmf_stat}
\alias{asap_nmf_stat}
\title{Topic statistics to estimate factor loading}
\usage{
asap_nmf_stat(
  y_dn,
  log_beta,
  beta_row_names,
  r_log_delta = NULL,
  do_stdize_beta = TRUE,
  do_log1p = FALSE,
  verbose = FALSE,
  NUM_THREADS = 1L,
  BLOCK_SIZE = 100L,
  MAX_ROW_WORD = 2L,
  ROW_WORD_SEP = "_",
  MAX_COL_WORD = 100L,
  COL_WORD_SEP = "@"
)
}
\arguments{
\item{y_dn}{sparse data matrix (D x N)}

\item{log_beta}{D x K log dictionary/design matrix}

\item{beta_row_names}{row names log_beta (D vector)}

\item{r_log_delta}{D x B log batch effect matrix}

\item{do_stdize_beta}{use standardized log_beta (Default: TRUE)}

\item{do_log1p}{do log(1+y) transformation}

\item{verbose}{verbosity}

\item{NUM_THREADS}{number of threads in data reading}

\item{BLOCK_SIZE}{disk I/O block size (number of columns)}

\item{MAX_ROW_WORD}{maximum words per line in \code{row_files[i]}}

\item{ROW_WORD_SEP}{word separation character to replace white space}

\item{MAX_COL_WORD}{maximum words per line in \code{col_files[i]}}

\item{COL_WORD_SEP}{word separation character to replace white space}
}
\value{
a list that contains:
\itemize{
\item beta dictionary matrix (row x factor)
\item corr empirical correlation (column x factor)
\item colsum the sum of each column (column x 1)
}
}
\description{
Topic statistics to estimate factor loading
}
