% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{asap_pmf_regression_mtx}
\alias{asap_pmf_regression_mtx}
\title{PMF regression}
\usage{
asap_pmf_regression_mtx(
  mtx_file,
  row_file,
  col_file,
  idx_file,
  log_beta,
  beta_row_names,
  log_delta = NULL,
  do_stdize_beta = TRUE,
  do_stdize_r = TRUE,
  do_log1p = FALSE,
  a0 = 1,
  b0 = 1,
  max_iter = 10L,
  verbose = FALSE,
  NUM_THREADS = 0L,
  BLOCK_SIZE = 1000L,
  MAX_ROW_WORD = 2L,
  ROW_WORD_SEP = "_",
  MAX_COL_WORD = 100L,
  COL_WORD_SEP = "@"
)
}
\arguments{
\item{mtx_file}{matrix-market-formatted data file (D x N, bgzip)}

\item{row_file}{row names file (D x 1)}

\item{col_file}{column names file (N x 1)}

\item{idx_file}{matrix-market colum index file}

\item{log_beta}{D x K log dictionary/design matrix (default: TRUE)}

\item{beta_row_names}{row names log_beta (D vector)}

\item{log_delta}{D x B log batch effect matrix}

\item{do_stdize_beta}{use standardized log_beta (Default: FALSE)}

\item{do_stdize_r}{standardize correlation matrix R (default: TRUE)}

\item{do_log1p}{do log(1+y) transformation}

\item{a0}{gamma(a0, b0) (default: 1)}

\item{b0}{gamma(a0, b0) (default: 1)}

\item{max_iter}{maximum iterations (default: 10)}

\item{verbose}{verbosity}

\item{NUM_THREADS}{number of threads in data reading}

\item{BLOCK_SIZE}{disk I/O block size (number of columns)}

\item{MAX_ROW_WORD}{maximum words per line in \code{row_files[i]}}

\item{ROW_WORD_SEP}{word separation character to replace white space}

\item{MAX_COL_WORD}{maximum words per line in \code{col_files[i]}}

\item{COL_WORD_SEP}{word separation character to replace white space}
}
\value{
a list that contains:
\itemize{
\item beta dictionary matrix (row x factor)
\item delta the dictionary matrix of batch effects (row x batch)
\item corr empirical correlation (column x factor)
\item theta factor loading (column x factor)
\item log.theta log-scaled factor loading (column x factor)
\item colsum the sum of each column (column x 1)
\item rownames row/feature names
\item colnames column/sample names
}
}
\description{
PMF regression
}
