% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{asap_pmf_stat}
\alias{asap_pmf_stat}
\title{PMF statistics to estimate factor loading}
\usage{
asap_pmf_stat(
  y_dn,
  log_beta,
  beta_row_names,
  r_log_delta = NULL,
  do_stdize_beta = FALSE,
  do_stdize_r = FALSE,
  do_log1p = FALSE,
  verbose = FALSE,
  NUM_THREADS = 0L,
  CELL_NORM = 10000,
  BLOCK_SIZE = 1000L
)
}
\arguments{
\item{y_dn}{sparse data matrix (D x N)}

\item{log_beta}{D x K log dictionary/design matrix}

\item{beta_row_names}{row names log_beta (D vector)}

\item{r_log_delta}{D x B log batch effect matrix}

\item{do_stdize_beta}{use standardized log_beta (Default: FALSE)}

\item{do_stdize_r}{standardize correlation matrix R (default: FALSE)}

\item{do_log1p}{do log(1+y) transformation}

\item{verbose}{verbosity}

\item{NUM_THREADS}{number of threads in data reading}

\item{CELL_NORM}{sample normalization constant (default: 1e4)}

\item{BLOCK_SIZE}{disk I/O block size (number of columns)}

\item{MAX_ROW_WORD}{maximum words per line in \code{row_files[i]}}

\item{ROW_WORD_SEP}{word separation character to replace white space}

\item{MAX_COL_WORD}{maximum words per line in \code{col_files[i]}}

\item{COL_WORD_SEP}{word separation character to replace white space}
}
\value{
a list that contains:
\itemize{
\item beta dictionary matrix (row x factor)
\item delta the dictionary matrix of batch effects (row x batch)
\item corr empirical correlation (column x factor)
\item colsum the sum of each column (column x 1)
\item rownames row names
}
}
\description{
PMF statistics to estimate factor loading
}
