% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{asap_pmf_stat_cbind_mtx}
\alias{asap_pmf_stat_cbind_mtx}
\title{PMF statistics to estimate factor loading}
\usage{
asap_pmf_stat_cbind_mtx(
  mtx_files,
  row_files,
  col_files,
  idx_files,
  log_beta,
  beta_row_names,
  log_delta = NULL,
  r_batch_names = NULL,
  rename_columns = FALSE,
  do_stdize_beta = TRUE,
  do_stdize_r = TRUE,
  do_log1p = FALSE,
  verbose = FALSE,
  a0 = 1,
  b0 = 1,
  max_iter = 10L,
  NUM_THREADS = 0L,
  BLOCK_SIZE = 1000L,
  MAX_ROW_WORD = 2L,
  ROW_WORD_SEP = "_",
  MAX_COL_WORD = 100L,
  COL_WORD_SEP = "@"
)
}
\arguments{
\item{log_beta}{D x K log dictionary/design matrix}

\item{beta_row_names}{row names log_beta (D vector)}

\item{log_delta}{D x B log batch effects}

\item{r_batch_names}{batch names (optional)}

\item{rename_columns}{append batch name at the end of each column name (default: FALSE)}

\item{do_stdize_beta}{use standardized log_beta (default: TRUE)}

\item{do_stdize_r}{standardize correlation matrix R (default: TRUE)}

\item{do_log1p}{do log(1+y) transformation (default: FALSE)}

\item{verbose}{verbosity}

\item{a0}{gamma(a0, b0) (default: 1)}

\item{b0}{gamma(a0, b0) (default: 1)}

\item{max_iter}{maximum iterations (default: 10)}

\item{NUM_THREADS}{number of threads in data reading}

\item{BLOCK_SIZE}{disk I/O block size (number of columns)}

\item{MAX_ROW_WORD}{maximum words per line in \code{row_files[i]}}

\item{ROW_WORD_SEP}{word separation character to replace white space}

\item{MAX_COL_WORD}{maximum words per line in \code{col_files[i]}}

\item{COL_WORD_SEP}{word separation character to replace white space}

\item{mtx_file}{matrix-market-formatted data file (D x N, bgzip)}

\item{row_file}{row names file (D x 1)}

\item{col_file}{column names file (N x 1)}

\item{idx_file}{matrix-market colum index file}
}
\value{
a list that contains:
\itemize{
\item beta the dictionary matrix (row x factor)
\item corr empirical correlation matrices (column x factor)
\item colsum column sum (column x 1)
\item rownames row names
\item batch.names batch names (based on
\item batch.index
\item colnames column names
}
}
\description{
PMF statistics to estimate factor loading
}
