% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{asap_pmf_stat_mtx}
\alias{asap_pmf_stat_mtx}
\title{PMF statistics to estimate factor loading}
\usage{
asap_pmf_stat_mtx(
  mtx_file,
  row_file,
  col_file,
  idx_file,
  log_beta,
  beta_row_names,
  r_log_delta = NULL,
  do_stdize_beta = TRUE,
  do_log1p = FALSE,
  verbose = FALSE,
  NUM_THREADS = 0L,
  CELL_NORM = 10000,
  BLOCK_SIZE = 1000L,
  MAX_ROW_WORD = 2L,
  ROW_WORD_SEP = "_",
  MAX_COL_WORD = 100L,
  COL_WORD_SEP = "@"
)
}
\arguments{
\item{mtx_file}{matrix-market-formatted data file (D x N, bgzip)}

\item{row_file}{row names file (D x 1)}

\item{col_file}{column names file (N x 1)}

\item{idx_file}{matrix-market colum index file}

\item{log_beta}{D x K log dictionary/design matrix}

\item{beta_row_names}{row names log_beta (D vector)}

\item{r_log_delta}{D x B log batch effect matrix}

\item{do_stdize_beta}{use standardized log_beta (Default: TRUE)}

\item{do_log1p}{do log(1+y) transformation}

\item{verbose}{verbosity}

\item{NUM_THREADS}{number of threads in data reading}

\item{CELL_NORM}{sample normalization constant (default: 1e4)}

\item{BLOCK_SIZE}{disk I/O block size (number of columns)}

\item{MAX_ROW_WORD}{maximum words per line in \code{row_files[i]}}

\item{ROW_WORD_SEP}{word separation character to replace white space}

\item{MAX_COL_WORD}{maximum words per line in \code{col_files[i]}}

\item{COL_WORD_SEP}{word separation character to replace white space}
}
\value{
a list that contains:
\itemize{
\item beta the dictionary matrix of topics (row x factor)
\item delta the dictionary matrix of batch effects (row x batch)
\item corr empirical correlation (column x factor)
\item colsum the sum of each column (column x 1)
\item rownames row names
\item rownames column names
}
}
\description{
PMF statistics to estimate factor loading
}
