% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{asap_pmf_stat_rbind}
\alias{asap_pmf_stat_rbind}
\title{Topic statistics to estimate factor loading}
\usage{
asap_pmf_stat_rbind(
  mtx_files,
  row_files,
  col_files,
  idx_files,
  log_beta_vec,
  beta_row_names_vec,
  do_stdize_beta = FALSE,
  do_log1p = FALSE,
  verbose = FALSE,
  NUM_THREADS = 1L,
  BLOCK_SIZE = 100L,
  MAX_ROW_WORD = 2L,
  ROW_WORD_SEP = "_",
  MAX_COL_WORD = 100L,
  COL_WORD_SEP = "@"
)
}
\arguments{
\item{do_stdize_beta}{use standardized log_beta (default: TRUE)}

\item{do_log1p}{do log(1+y) transformation (default: FALSE)}

\item{verbose}{verbosity}

\item{NUM_THREADS}{number of threads in data reading}

\item{BLOCK_SIZE}{disk I/O block size (number of columns)}

\item{MAX_ROW_WORD}{maximum words per line in \code{row_files[i]}}

\item{ROW_WORD_SEP}{word separation character to replace white space}

\item{MAX_COL_WORD}{maximum words per line in \code{col_files[i]}}

\item{COL_WORD_SEP}{word separation character to replace white space}

\item{mtx_file}{matrix-market-formatted data file (D x N, bgzip)}

\item{row_file}{row names file (D x 1)}

\item{col_file}{column names file (N x 1)}

\item{idx_file}{matrix-market colum index file}

\item{log_x}{D x K log dictionary/design matrix}

\item{beta_row_names}{row names log_x (D vector)}
}
\value{
a list that contains:
\itemize{
\item beta.list a list of dictionary matrices (row x factor)
\item corr.list a list of empirical correlation matrices (column x factor)
\item colsum.list a list of column sum vectors (column x 1)
}
}
\description{
Topic statistics to estimate factor loading
}
