% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{asap_predict_mtx}
\alias{asap_predict_mtx}
\title{Predict NMF loading -- this may be slow for high-dim data}
\usage{
asap_predict_mtx(
  mtx_file,
  memory_location,
  beta_dict,
  do_beta_rescale = FALSE,
  collapsing = NULL,
  mcem = 100L,
  burnin = 10L,
  latent_iter = 10L,
  thining = 3L,
  a0 = 1,
  b0 = 1,
  rseed = 42L,
  verbose = FALSE,
  NUM_THREADS = 1L,
  BLOCK_SIZE = 100L,
  gibbs_sampling = FALSE
)
}
\arguments{
\item{mtx_file}{matrix-market-formatted data file (bgzip)}

\item{memory_location}{column indexing for the mtx}

\item{beta_dict}{row x factor dictionary (beta) matrix}

\item{do_beta_rescale}{rescale the columns of the beta matrix}

\item{collapsing}{r x row collapsing matrix (r < row)}

\item{mcem}{number of Monte Carlo Expectation Maximization}

\item{burnin}{burn-in period}

\item{latent_iter}{latent sampling steps}

\item{thining}{thining interval in record keeping}

\item{a0}{gamma(a0, b0)}

\item{b0}{gamma(a0, b0)}

\item{rseed}{random seed}

\item{verbose}{verbosity}

\item{NUM_THREADS}{number of threads in data reading}

\item{BLOCK_SIZE}{disk I/O block size (number of columns)}
}
\description{
Predict NMF loading -- this may be slow for high-dim data
}
