% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{asap_random_bulk}
\alias{asap_random_bulk}
\title{Generate approximate pseudo-bulk data by random projections}
\usage{
asap_random_bulk(
  mtx_file,
  row_file,
  col_file,
  idx_file,
  num_factors,
  r_covar_n = NULL,
  r_covar_d = NULL,
  rseed = 42L,
  verbose = FALSE,
  NUM_THREADS = 1L,
  BLOCK_SIZE = 100L,
  do_log1p = FALSE,
  do_down_sample = FALSE,
  save_rand_proj = FALSE,
  weighted_rand_proj = FALSE,
  CELL_PER_SAMPLE = 100L,
  a0 = 1e-08,
  b0 = 1,
  MAX_ROW_WORD = 2L,
  ROW_WORD_SEP = "_",
  MAX_COL_WORD = 100L,
  COL_WORD_SEP = "@"
)
}
\arguments{
\item{mtx_file}{matrix-market-formatted data file (bgzip)}

\item{row_file}{row names (gene/feature names)}

\item{col_file}{column names (cell/column names)}

\item{idx_file}{matrix-market colum index file}

\item{num_factors}{a desired number of random factors}

\item{r_covar_n}{N x r covariates (default: NULL)}

\item{r_covar_d}{D x r covariates (default: NULL)}

\item{rseed}{random seed}

\item{verbose}{verbosity}

\item{NUM_THREADS}{number of threads in data reading}

\item{BLOCK_SIZE}{disk I/O block size (number of columns)}

\item{do_log1p}{log(x + 1) transformation (default: FALSE)}

\item{do_down_sample}{down-sampling (default: FALSE)}

\item{save_rand_proj}{save random projection (default: FALSE)}

\item{weighted_rand_proj}{save random projection (default: FALSE)}

\item{CELL_PER_SAMPLE}{down-sampling cell per sample (default: 100)}

\item{a0}{gamma(a0, b0) (default: 1e-8)}

\item{b0}{gamma(a0, b0) (default: 1)}

\item{MAX_ROW_WORD}{maximum words per line in \code{row_file}}

\item{ROW_WORD_SEP}{word separation character to replace white space}

\item{MAX_COL_WORD}{maximum words per line in \code{col_file}}

\item{COL_WORD_SEP}{word separation character to replace white space}
}
\value{
a list
\itemize{
\item \code{PB} pseudobulk (average) data (feature x sample)
\item \code{sum} pseudobulk (sum) data (feature x sample)
\item \code{size} size per sample (sample x 1)
\item \code{positions} pseudobulk sample positions (cell x 1)
\item \code{rand.dict} random dictionary (proj factor x feature)
\item \code{rand.proj} random projection results (sample x proj factor)
\item \code{colnames} column (cell) names
\item \code{rownames} feature (gene) names
}
}
\description{
Generate approximate pseudo-bulk data by random projections
}
