% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{asap_random_bulk_cbind}
\alias{asap_random_bulk_cbind}
\title{Generate approximate pseudo-bulk data by random projections
while sharing rows/features across multiple data sets.
Horizontal concatenation.}
\usage{
asap_random_bulk_cbind(
  y_dn_vec,
  num_factors,
  r_row_names = NULL,
  r_batch_names = NULL,
  rseed = 42L,
  verbose = TRUE,
  NUM_THREADS = 1L,
  BLOCK_SIZE = 100L,
  do_batch_adj = TRUE,
  do_log1p = FALSE,
  do_down_sample = TRUE,
  save_aux_data = FALSE,
  KNN_CELL = 10L,
  CELL_PER_SAMPLE = 100L,
  BATCH_ADJ_ITER = 100L,
  a0 = 1,
  b0 = 1
)
}
\arguments{
\item{y_dn_vec}{a list of sparse matrices}

\item{num_factors}{a desired number of random factors}

\item{rseed}{random seed}

\item{verbose}{verbosity}

\item{NUM_THREADS}{number of threads in data reading}

\item{BLOCK_SIZE}{disk I/O block size (number of columns)}

\item{do_batch_adj}{(default: FALSE)}

\item{do_log1p}{log(x + 1) transformation (default: FALSE)}

\item{do_down_sample}{down-sampling (default: TRUE)}

\item{save_aux_data}{save auxiliary data (default: FALSE)}

\item{KNN_CELL}{k-NN cells per batch between different batches (default: 10)}

\item{CELL_PER_SAMPLE}{down-sampling cell per sample (default: 100)}

\item{BATCH_ADJ_ITER}{batch Adjustment steps (default: 100)}

\item{a0}{gamma(a0, b0) (default: 1e-8)}

\item{b0}{gamma(a0, b0) (default: 1)}

\item{take_union_rows}{take union of rows (default: FALSE)}

\item{weighted_rand_proj}{save random projection (default: FALSE)}

\item{MAX_ROW_WORD}{maximum words per line in \code{row_files[i]}}

\item{ROW_WORD_SEP}{word separation character to replace white space}

\item{MAX_COL_WORD}{maximum words per line in \code{col_files[i]}}

\item{COL_WORD_SEP}{word separation character to replace white space}
}
\value{
a list
\itemize{
\item \code{PB} pseudobulk (average) data (feature x sample)
\item \code{sum} pseudobulk (sum) data (feature x sample)
\item \code{matched.sum} kNN-matched pseudobulk data (feature x sample)
\item \code{sum_db} batch-specific sum (feature x batch)
\item \code{size} size per sample (sample x 1)
\item \code{prob_bs} batch-specific frequency (batch x sample)
\item \code{size_bs} batch-specific size (batch x sample)
\item \code{batch.effect} batch effect (feature x batch)
\item \code{log.batch.effect} log batch effect (feature x batch)
\item \code{batch.names} batch names (batch x 1)
\item \code{positions} pseudobulk sample positions (cell x 1)
\item \code{rand.dict} random dictionary (proj factor x feature)
\item \code{rand.proj} random projection results (sample x proj factor)
\item \code{colnames} column (cell) names
\item \code{rownames} feature (gene) names
}
}
\description{
Generate approximate pseudo-bulk data by random projections
while sharing rows/features across multiple data sets.
Horizontal concatenation.
}
