% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{asap_random_bulk_data}
\alias{asap_random_bulk_data}
\title{Generate approximate pseudo-bulk data by random projections}
\usage{
asap_random_bulk_data(
  mtx_file,
  memory_location,
  num_factors,
  rseed = 42L,
  verbose = FALSE,
  NUM_THREADS = 1L,
  BLOCK_SIZE = 100L
)
}
\arguments{
\item{mtx_file}{matrix-market-formatted data file (bgzip)}

\item{memory_location}{column indexing for the mtx}

\item{num_factors}{a desired number of random factors}

\item{rseed}{random seed}

\item{verbose}{verbosity}

\item{NUM_THREADS}{number of threads in data reading}

\item{BLOCK_SIZE}{disk I/O block size (number of columns)}
}
\description{
Generate approximate pseudo-bulk data by random projections
}
