% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{asap_random_bulk_interacting_columns}
\alias{asap_random_bulk_interacting_columns}
\title{Generate approximate pseudo-bulk interaction data by random projections}
\usage{
asap_random_bulk_interacting_columns(
  mtx_file,
  row_file,
  col_file,
  idx_file,
  num_factors,
  W_nm_list,
  mtx_file_rhs = NULL,
  row_file_rhs = NULL,
  col_file_rhs = NULL,
  idx_file_rhs = NULL,
  A_dd_list = NULL,
  rseed = 42L,
  do_product = FALSE,
  do_log1p = FALSE,
  do_down_sample = FALSE,
  save_rand_proj = FALSE,
  weighted_rand_proj = FALSE,
  NUM_THREADS = 0L,
  CELL_NORM = 10000,
  BLOCK_SIZE = 1000L,
  EDGE_PER_SAMPLE = 100L,
  a0 = 1,
  b0 = 1,
  MAX_ROW_WORD = 2L,
  ROW_WORD_SEP = "_",
  MAX_COL_WORD = 100L,
  COL_WORD_SEP = "@",
  verbose = FALSE
)
}
\arguments{
\item{mtx_file}{matrix-market-formatted data file (bgzip)}

\item{row_file}{row names (gene/feature names)}

\item{col_file}{column names (cell/column names)}

\item{idx_file}{matrix-market colum index file}

\item{num_factors}{a desired number of random factors}

\item{W_nm_list}{list(src.index, tgt.index, \link{weights}) for columns}

\item{mtx_file_rhs}{right-hand-side matrix-market-formatted data file (bgzip)}

\item{row_file_rhs}{right-hand-side row names (gene/feature names)}

\item{col_file_rhs}{right-hand-side column names (cell/column names)}

\item{idx_file_rhs}{right-hand-side matrix-market colum index file}

\item{A_dd_list}{list(src.index, tgt.index, \link{weights}) for features}

\item{rseed}{random seed}

\item{do_product}{yi * yj for interaction (default: FALSE)}

\item{do_log1p}{log(x + 1) transformation (default: FALSE)}

\item{do_down_sample}{down-sampling (default: FALSE)}

\item{save_rand_proj}{save random projection (default: FALSE)}

\item{weighted_rand_proj}{save random projection (default: FALSE)}

\item{NUM_THREADS}{number of threads in data reading}

\item{CELL_NORM}{normalization constant per each data point}

\item{BLOCK_SIZE}{disk I/O block size (number of columns)}

\item{EDGE_PER_SAMPLE}{down-sampling cell per sample (default: 100)}

\item{a0}{gamma(a0, b0) (default: 1)}

\item{b0}{gamma(a0, b0) (default: 1)}

\item{MAX_ROW_WORD}{maximum words per line in \code{row_file}}

\item{ROW_WORD_SEP}{word separation character to replace white space}

\item{MAX_COL_WORD}{maximum words per line in \code{col_file}}

\item{COL_WORD_SEP}{word separation character to replace white space}

\item{verbose}{verbosity}
}
\value{
a list
\itemize{
\item \code{PB} pseudobulk (average) data (feature x sample)
\item \code{sum} pseudobulk (sum) data (feature x sample)
\item \code{size} size per sample (sample x 1)
\item \code{positions} pseudobulk sample positions (cell pair x 1)
\item \code{rand.dict} random dictionary (proj factor x feature)
\item \code{rand.proj} random projection results (sample x proj factor)
\item \code{colnames} column (cell) names
\item \code{rownames} feature (gene) names
}
}
\description{
Generate approximate pseudo-bulk interaction data by random projections
}
