% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{asap_random_bulk_multi_hcat}
\alias{asap_random_bulk_multi_hcat}
\title{Generate approximate pseudo-bulk data by random projections
while sharing rows/features across multiple data sets.
Horizontal concatenation.}
\usage{
asap_random_bulk_multi_hcat(
  mtx_files,
  row_files,
  col_files,
  idx_files,
  num_factors,
  take_union_rows = FALSE,
  rseed = 42L,
  verbose = TRUE,
  NUM_THREADS = 1L,
  BLOCK_SIZE = 100L,
  do_batch_adj = TRUE,
  do_log1p = FALSE,
  do_down_sample = TRUE,
  save_rand_proj = FALSE,
  KNN_CELL = 3L,
  CELL_PER_SAMPLE = 100L,
  BATCH_ADJ_ITER = 100L,
  a0 = 1e-08,
  b0 = 1,
  MAX_ROW_WORD = 2L,
  ROW_WORD_SEP = "_",
  MAX_COL_WORD = 100L,
  COL_WORD_SEP = "@"
)
}
\arguments{
\item{mtx_files}{matrix-market-formatted data files (bgzip)}

\item{row_files}{row names (gene/feature names)}

\item{col_files}{column names (cell/column names)}

\item{idx_files}{matrix-market colum index files}

\item{num_factors}{a desired number of random factors}

\item{take_union_rows}{take union of rows (default: FALSE)}

\item{rseed}{random seed}

\item{verbose}{verbosity}

\item{NUM_THREADS}{number of threads in data reading}

\item{BLOCK_SIZE}{disk I/O block size (number of columns)}

\item{do_batch_adj}{(default: FALSE)}

\item{do_log1p}{log(x + 1) transformation (default: FALSE)}

\item{do_down_sample}{down-sampling (default: TRUE)}

\item{save_rand_proj}{save random projection (default: FALSE)}

\item{KNN_CELL}{k-NN cells per batch between different batches (default: 3)}

\item{CELL_PER_SAMPLE}{down-sampling cell per sample (default: 100)}

\item{BATCH_ADJ_ITER}{batch Adjustment steps (default: 100)}

\item{a0}{gamma(a0, b0) (default: 1e-8)}

\item{b0}{gamma(a0, b0) (default: 1)}

\item{MAX_ROW_WORD}{maximum words per line in \code{row_files[i]}}

\item{ROW_WORD_SEP}{word separation character to replace white space}

\item{MAX_COL_WORD}{maximum words per line in \code{col_files[i]}}

\item{COL_WORD_SEP}{word separation character to replace white space}
}
\value{
a list
\itemize{
\item \code{PB} pseudobulk (average) data (feature x sample)
\item \code{sum} pseudobulk (sum) data (feature x sample)
\item \code{matched.sum} kNN-matched pseudobulk data (feature x sample)
\item \code{sum_db} batch-specific sum (feature x batch)
\item \code{size} size per sample (sample x 1)
\item \code{prob_bs} batch-specific frequency (batch x sample)
\item \code{size_bs} batch-specific size (batch x sample)
\item \code{batch.effect} batch effect (feature x batch)
\item \code{log.batch.effect} log batch effect (feature x batch)
\item \code{batch.names} batch names (batch x 1)
\item \code{positions} pseudobulk sample positions (cell x 1)
\item \code{rand.dict} random dictionary (proj factor x feature)
\item \code{rand.proj} random projection results (sample x proj factor)
\item \code{colnames} column (cell) names
\item \code{rownames} feature (gene) names
}
}
\description{
Generate approximate pseudo-bulk data by random projections
while sharing rows/features across multiple data sets.
Horizontal concatenation.
}
