% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{asap_random_bulk_rbind}
\alias{asap_random_bulk_rbind}
\title{Generate approximate pseudo-bulk data by random projections
while sharing columns/cells across multiple data sets.
Vertical concatenation.}
\usage{
asap_random_bulk_rbind(
  mtx_files,
  row_files,
  col_files,
  idx_files,
  num_factors,
  rseed = 42L,
  verbose = TRUE,
  NUM_THREADS = 1L,
  BLOCK_SIZE = 100L,
  do_log1p = FALSE,
  do_down_sample = FALSE,
  save_rand_proj = FALSE,
  weighted_rand_proj = FALSE,
  CELL_PER_SAMPLE = 100L,
  a0 = 1e-08,
  b0 = 1,
  MAX_ROW_WORD = 2L,
  ROW_WORD_SEP = "_",
  MAX_COL_WORD = 100L,
  COL_WORD_SEP = "@"
)
}
\arguments{
\item{mtx_files}{matrix-market-formatted data files (bgzip)}

\item{row_files}{row names (gene/feature names)}

\item{col_files}{column names (cell/column names)}

\item{idx_files}{matrix-market colum index files}

\item{num_factors}{a desired number of random factors per data set}

\item{rseed}{random seed}

\item{verbose}{verbosity}

\item{NUM_THREADS}{number of threads in data reading}

\item{BLOCK_SIZE}{disk I/O block size (number of columns)}

\item{do_log1p}{log(x + 1) transformation (default: FALSE)}

\item{do_down_sample}{down-sampling (default: FALSE)}

\item{save_rand_proj}{save random projection (default: FALSE)}

\item{weighted_rand_proj}{save random projection (default: FALSE)}

\item{CELL_PER_SAMPLE}{down-sampling cell per sample (default: 100)}

\item{a0}{gamma(a0, b0) (default: 1e-8)}

\item{b0}{gamma(a0, b0) (default: 1)}

\item{MAX_ROW_WORD}{maximum words per line in \code{row_files[i]}}

\item{ROW_WORD_SEP}{word separation character to replace white space}

\item{MAX_COL_WORD}{maximum words per line in \code{col_files[i]}}

\item{COL_WORD_SEP}{word separation character to replace white space}
}
\value{
a list
\itemize{
\item \code{PB.list} pseudobulk (average) data (feature x sample) for each type
\item \code{sum.list} pseudobulk (sum) data (feature x sample) for each type
\item \code{size.list} size per sample (sample x 1) for each type
\item \code{rownames.list} feature (gene) names for each type
\item \code{colnames} column (cell) names across data types
\item \code{positions} pseudobulk sample positions (cell x 1)
\item \code{rand.proj} random projection results (sample x proj factor)
\item \code{colnames} column (cell) names
}
}
\description{
Generate approximate pseudo-bulk data by random projections
while sharing columns/cells across multiple data sets.
Vertical concatenation.
}
