% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{asap_regression}
\alias{asap_regression}
\title{Poisson regression to estimate factor loading}
\usage{
asap_regression(
  Y_,
  log_x,
  a0 = 1e-08,
  b0 = 1,
  max_iter = 10L,
  do_log1p = FALSE,
  verbose = TRUE
)
}
\arguments{
\item{log_x}{D x K log dictionary/design matrix}

\item{a0}{gamma(a0, b0) (default: 1e-8)}

\item{b0}{gamma(a0, b0) (default: 1)}

\item{do_log1p}{do log(1+y) transformation (default: FALSE)}

\item{verbose}{verbosity (default: false)}

\item{Y}{D x N data matrix}
}
\description{
Poisson regression to estimate factor loading
}
