% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{asap_regression_mtx}
\alias{asap_regression_mtx}
\title{Poisson regression to estimate factor loading}
\usage{
asap_regression_mtx(
  mtx_file,
  memory_location,
  log_x,
  a0 = 1,
  b0 = 1,
  max_iter = 10L,
  verbose = FALSE,
  NUM_THREADS = 1L,
  BLOCK_SIZE = 100L
)
}
\arguments{
\item{mtx_file}{matrix-market-formatted data file (bgzip)}

\item{memory_location}{column indexing for the mtx}

\item{log_x}{D x K log dictionary/design matrix}

\item{a0}{gamma(a0, b0)}

\item{b0}{gamma(a0, b0)}

\item{verbose}{verbosity}

\item{NUM_THREADS}{number of threads in data reading}

\item{BLOCK_SIZE}{disk I/O block size (number of columns)}
}
\description{
Poisson regression to estimate factor loading
}
