% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{asap_stretch_nn_matrix_columns}
\alias{asap_stretch_nn_matrix_columns}
\title{Stretch non-negative matrix}
\usage{
asap_stretch_nn_matrix_columns(
  Y,
  qq_min = 0.01,
  qq_max = 0.99,
  std_min = -8,
  std_max = 8,
  verbose = TRUE
)
}
\arguments{
\item{Y}{non-negative data matrix(gene x sample)}

\item{qq_min}{min quantile (default: 0.01)}

\item{qq_max}{min quantile (default: 0.99)}

\item{std_min}{min after standardization of log (default: -8)}

\item{std_max}{max after standardization of log (default: 8)}

\item{verbose}{speak more (default: TRUE)}
}
\description{
Stretch non-negative matrix
}
