% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{asap_topic_pmf}
\alias{asap_topic_pmf}
\title{Calibrate topic proportions based on sufficient statistics}
\usage{
asap_topic_pmf(
  beta_dk,
  R_nk,
  Y_n,
  a0 = 1,
  b0 = 1,
  max_iter = 10L,
  NUM_THREADS = 0L,
  stdize_r = TRUE,
  verbose = TRUE
)
}
\arguments{
\item{beta_dk}{dictionary matrix (feature D  x factor K)}

\item{R_nk}{correlation matrix (sample N x factor K)}

\item{Y_n}{sum vector (sample N x 1)}

\item{a0}{gamma(a0, b0) (default: 1)}

\item{b0}{gamma(a0, b0) (default: 1)}

\item{max_iter}{maximum iterations (default: 10)}

\item{NUM_THREADS}{number of parallel threads (default: 1)}

\item{stdize_r}{standardize correlation matrix R (default: TRUE)}

\item{verbose}{(default: TRUE)}
}
\value{
a list that contains:
\itemize{
\item beta (D x K) matrix
\item theta (N x K) matrix
\item log.theta (N x K) log matrix
\item log.theta.sd (N x K) standard deviation matrix
}
}
\description{
Calibrate topic proportions based on sufficient statistics
}
