% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{asap_topic_stat}
\alias{asap_topic_stat}
\title{Topic statistics to estimate factor loading}
\usage{
asap_topic_stat(
  mtx_file,
  row_file,
  col_file,
  idx_file,
  log_beta,
  x_row_names,
  do_log1p = FALSE,
  verbose = FALSE,
  NUM_THREADS = 1L,
  BLOCK_SIZE = 100L,
  MAX_ROW_WORD = 2L,
  ROW_WORD_SEP = "_",
  MAX_COL_WORD = 100L,
  COL_WORD_SEP = "@"
)
}
\arguments{
\item{mtx_file}{matrix-market-formatted data file (D x N, bgzip)}

\item{row_file}{row names file (D x 1)}

\item{col_file}{column names file (N x 1)}

\item{idx_file}{matrix-market colum index file}

\item{log_beta}{D x K log dictionary/design matrix}

\item{x_row_names}{row names log_beta (D vector)}

\item{do_log1p}{do log(1+y) transformation}

\item{verbose}{verbosity}

\item{NUM_THREADS}{number of threads in data reading}

\item{BLOCK_SIZE}{disk I/O block size (number of columns)}

\item{MAX_ROW_WORD}{maximum words per line in \code{row_files[i]}}

\item{ROW_WORD_SEP}{word separation character to replace white space}

\item{MAX_COL_WORD}{maximum words per line in \code{col_files[i]}}

\item{COL_WORD_SEP}{word separation character to replace white space}
}
\value{
a list that contains:
\itemize{
\item beta dictionary matrix (row x factor)
\item corr empirical correlation (column x factor)
\item colsum the sum of each column (column x 1)
}
}
\description{
Topic statistics to estimate factor loading
}
