% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{collapse_network}
\alias{collapse_network}
\title{Collapse N x N adjacency network into S x S}
\usage{
collapse_network(W_nn_list, r_positions, N, S, verbose = TRUE)
}
\arguments{
\item{W_nn_list}{adjacency list}

\item{r_positions}{collapsing positions}

\item{N}{number of vertices}

\item{S}{number of meta-vertices}

\item{verbose}{verbosity}
}
\value{
a list of {rows, columns, weights}
}
\description{
Collapse N x N adjacency network into S x S
}
