% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{decompose_network}
\alias{decompose_network}
\title{Assign best topic membership for the edges}
\usage{
decompose_network(A_dd, beta_dt, cutoff = 1e-08, verbose = TRUE)
}
\arguments{
\item{A_dd}{D x D adjacency matrix}

\item{beta_dt}{D x T node propensity matrix}

\item{cutoff}{A\link{i,j} cutoff (default: 1e-8)}

\item{verbose}{verbosity}
}
\description{
Assign best topic membership for the edges
}
