% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Methods.R
\name{fit.topic.asap}
\alias{fit.topic.asap}
\title{Fit a topic model by ASAP}
\usage{
fit.topic.asap(
  mtx.file,
  k,
  pb.factors = k,
  num.proj = 3,
  em.step = 100,
  e.step = 1,
  deg.step = 0,
  .burnin = 10,
  .thining = 3,
  do.collapse.rows = TRUE,
  .collapse.discrete = FALSE,
  .collapsing.level = 100,
  .collapsing.dpm = 1,
  .collapsing.mcmc = 500,
  a0 = 0.01,
  b0 = 1e-04,
  index.file = paste0(mtx.file, ".index"),
  verbose = TRUE,
  num.threads = 1,
  block.size = 100,
  eval.llik = FALSE,
  .rand.seed = 42,
  .sample.col.row = FALSE
)
}
\arguments{
\item{mtx.file}{data matrix file in a matrix market format}

\item{k}{number of topics}

\item{num.proj}{the number of pseudobulk projection steps}

\item{em.step}{Monte Carlo EM steps (default: 100)}

\item{e.step}{Num of E-step MCMC steps (default: 1)}

\item{deg.step}{Num of Degree calibration steps (default: 0)}

\item{.burnin}{burn-in period in the record keeping (default: 10)}

\item{.thining}{thining for the record keeping (default: 3)}

\item{do.collapse.rows}{collapse rows to speed up the final NMF}

\item{.collapsing.level}{collapsed dimension (default: 100)}

\item{.collapsing.dpm}{(default: 1)}

\item{.collapsing.mcmc}{(default: 500)}

\item{a0}{Gamma(a0, b0) prior (default: 1e-2)}

\item{b0}{Gamma(a0, b0) prior (default: 1e-4)}

\item{index.file}{a file for column indexes (default: "{mtx.file}.index")}

\item{verbose}{verbosity}

\item{num.threads}{number of threads (default: 1)}

\item{block.size}{a block size for disk I/O (default: 100)}

\item{eval.llik}{evaluate log-likelihood trace in PMF (default: FALSE)}

\item{.rand.seed}{random seed (default: 42)}

\item{.sample.col.row}{Take column-wise MH step (default: TRUE)}

\item{.collapsing.discrete}{do the row collapsing after discretization}
}
\description{
Fit a topic model by ASAP
}
