% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{mmutil_copy_selected_columns}
\alias{mmutil_copy_selected_columns}
\title{Take a subset of columns and create a new MTX file-set}
\usage{
mmutil_copy_selected_columns(
  mtx_file,
  row_file,
  col_file,
  r_selected,
  output,
  MAX_COL_WORD = 100L,
  COL_WORD_SEP = "@"
)
}
\arguments{
\item{mtx_file}{data file}

\item{row_file}{row file}

\item{col_file}{column file}

\item{output}{output header}

\item{MAX_COL_WORD}{maximum words per line in \code{col_file}}

\item{COL_WORD_SEP}{word separation character to replace white space}

\item{selected}{selected column names}
}
\description{
Take a subset of columns and create a new MTX file-set
}
\examples{

options(stringsAsFactors=FALSE)
rr <- rgamma(20, 1, 1)
mm <- matrix(rgamma(10 * 2, 1, 1), 10, 2)
src.hdr <- "test_org"
src.files <- mmutil_simulate_poisson(mm, rr, src.hdr)
Y <- Matrix::readMM(src.files$mtx)
colnames(Y) <- read.table(src.files$col)$V1
print(Y)
sub.cols <- sort(read.table(src.files$col)$V1[sample(20,3)])
print(sub.cols)
tgt.hdr <- "test_sub"
tgt.files <- mmutil_copy_selected_columns(
                         src.files$mtx,
                         src.files$row,
                         src.files$col,
                         sub.cols, tgt.hdr)
Y <- Matrix::readMM(tgt.files$mtx)
colnames(Y) <- read.table(tgt.files$col)$V1
print(Y)
unlink(list.files(pattern = src.hdr))
unlink(list.files(pattern = tgt.hdr))

}
