% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{mmutil_info}
\alias{mmutil_info}
\title{Just read the header information}
\usage{
mmutil_info(mtx_file)
}
\arguments{
\item{mtx_file}{data file}
}
\value{
info
}
\description{
Just read the header information
}
