% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{mmutil_read_index}
\alias{mmutil_read_index}
\title{Read an index file to R}
\usage{
mmutil_read_index(index_file)
}
\arguments{
\item{index_file}{index file}
}
\value{
a vector column index (a vector of memory locations)
}
\description{
Read an index file to R
}
