% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{mmutil_read_rows_columns}
\alias{mmutil_read_rows_columns}
\title{Read a subset of rows and columns from the data matrix}
\usage{
mmutil_read_rows_columns(
  mtx_file,
  memory_location,
  r_row_index,
  r_column_index,
  verbose = FALSE
)
}
\arguments{
\item{mtx_file}{data file}

\item{memory_location}{column -> memory location}

\item{r_row_index}{row indexes to retrieve (1-based)}

\item{r_column_index}{column indexes to retrieve (1-based)}

\item{verbose}{verbosity}
}
\value{
a dense sub-matrix
}
\description{
Read a subset of rows and columns from the data matrix
}
