% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{mmutil_write_mtx}
\alias{mmutil_write_mtx}
\title{Write down sparse matrix to the disk}
\usage{
mmutil_write_mtx(X, mtx_file)
}
\arguments{
\item{X}{sparse matrix}

\item{mtx_file}{file name}
}
\value{
EXIT_SUCCESS or EXIT_FAILURE
}
\description{
Write down sparse matrix to the disk
}
