% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Methods.R
\name{pmf2topic}
\alias{pmf2topic}
\title{Convert Poisson Matrix Factorization to Multinomial Topic model.
The same idea was first coined by \code{fastTopics} paper.}
\usage{
pmf2topic(.beta, .theta, eps = 1e-08)
}
\arguments{
\item{.beta}{D x K dictionary matrix}

\item{.theta}{N x K sample-wise factor loading matrix}
}
\value{
a list of (beta, prop, depth), where the beta is a re-scaled dictionary matrix, each row of the prop matrix corresponds to a mixing proportion per column, and the depth parameter gauges a sequencing depth for each column
}
\description{
Convert Poisson Matrix Factorization to Multinomial Topic model.
The same idea was first coined by \code{fastTopics} paper.
}
\details{
E\link{Y|.beta, .theta} = \code{.beta \%*\% t(.theta)}
}
