% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Wrapper.R
\name{read.mtx.sparse}
\alias{read.mtx.sparse}
\title{A wrapper function to read a sparse submatrix}
\usage{
read.mtx.sparse(
  mtx.file,
  sub.cols = NULL,
  memory.idx = NULL,
  memory.idx.file = paste0(mtx.file, ".index"),
  verbose = FALSE,
  num.threads = 1
)
}
\arguments{
\item{mtx.file}{matrix market file}

\item{sub.cols}{column index (default: all)}

\item{memory.idx}{memory locations}

\item{memory.idx.file}{memory location file}

\item{verbose}{verbosity}
}
\value{
a sparse matrix
}
\description{
A wrapper function to read a sparse submatrix
}
