% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{asap_fit_nmf_alternate}
\alias{asap_fit_nmf_alternate}
\title{A quick NMF estimation based on alternating Poisson regressions}
\usage{
asap_fit_nmf_alternate(
  Y_dn,
  maxK,
  max_iter = 100L,
  burnin = 10L,
  verbose = TRUE,
  discretize = TRUE,
  a0 = 1,
  b0 = 1,
  rseed = 1337L,
  EPS = 1e-06,
  rate_m = 1,
  rate_v = 1,
  svd_init = TRUE
)
}
\arguments{
\item{Y_dn}{non-negative data matrix (gene x sample)}

\item{maxK}{maximum number of factors}

\item{max_iter}{max number of optimization steps}

\item{burnin}{number of initiation steps}

\item{verbose}{verbosity}

\item{a0}{gamma(a0, b0) default: a0 = 1}

\item{b0}{gamma(a0, b0) default: b0 = 1}

\item{rseed}{random seed (default: 1337)}

\item{min_iter}{min number of optimization steps}
}
\value{
a list that contains:
\itemize{
\item log.likelihood log-likelihood trace
\item beta dictionary (gene x factor)
\item log.beta log-dictionary (gene x factor)
\item theta loading (sample x factor)
\item log.theta log-loading (sample x factor)
\item log.phi auxiliary variables (gene x factor)
\item log.rho auxiliary variables (sample x factor)
}
}
\description{
A quick NMF estimation based on alternating Poisson regressions
}
