% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{asap_random_bulk_data}
\alias{asap_random_bulk_data}
\title{Generate approximate pseudo-bulk data by random projections}
\usage{
asap_random_bulk_data(
  mtx_file,
  mtx_idx_file,
  num_factors,
  r_covar = NULL,
  r_batch = NULL,
  rseed = 42L,
  verbose = FALSE,
  NUM_THREADS = 1L,
  BLOCK_SIZE = 100L,
  do_scale = FALSE,
  do_log1p = FALSE,
  do_row_std = FALSE,
  KNN_CELL = 10L,
  BATCH_ADJ_ITER = 100L,
  a0 = 1,
  b0 = 1
)
}
\arguments{
\item{mtx_file}{matrix-market-formatted data file (bgzip)}

\item{mtx_idx_file}{matrix-market colum index file}

\item{num_factors}{a desired number of random factors}

\item{r_covar}{covariates (default: NULL)}

\item{r_batch}{batch information (default: NULL)}

\item{rseed}{random seed}

\item{verbose}{verbosity}

\item{NUM_THREADS}{number of threads in data reading}

\item{BLOCK_SIZE}{disk I/O block size (number of columns)}

\item{do_scale}{scale each column by standard deviation (default: TRUE)}

\item{do_log1p}{log(x + 1) transformation (default: FALSE)}

\item{do_row_std}{rowwise standardization (default: FALSE)}

\item{KNN_CELL}{k-NN matching between cells (default: 10)}

\item{BATCH_ADJ_ITER}{batch Adjustment steps (default: 100)}

\item{a0}{gamma(a0, b0) (default: 1)}

\item{b0}{gamma(a0, b0) (default: 1)}
}
\description{
Generate approximate pseudo-bulk data by random projections
}
