% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{asap_random_bulk_data_multi}
\alias{asap_random_bulk_data_multi}
\title{Generate approximate pseudo-bulk data by random projections}
\usage{
asap_random_bulk_data_multi(
  mtx_files,
  row_files,
  idx_files,
  num_factors,
  rseed = 42L,
  verbose = FALSE,
  NUM_THREADS = 1L,
  BLOCK_SIZE = 100L,
  do_log1p = FALSE,
  KNN_CELL = 10L,
  BATCH_ADJ_ITER = 100L,
  a0 = 1,
  b0 = 1
)
}
\arguments{
\item{mtx_files}{matrix-market-formatted data files (bgzip)}

\item{row_files}{row names (gene/feature names)}

\item{rseed}{random seed}

\item{verbose}{verbosity}

\item{NUM_THREADS}{number of threads in data reading}

\item{BLOCK_SIZE}{disk I/O block size (number of columns)}

\item{do_log1p}{log(x + 1) transformation (default: FALSE)}

\item{KNN_CELL}{k-NN matching between cells (default: 10)}

\item{BATCH_ADJ_ITER}{batch Adjustment steps (default: 100)}

\item{a0}{gamma(a0, b0) (default: 1)}

\item{b0}{gamma(a0, b0) (default: 1)}

\item{mtx_idx_files}{matrix-market colum index files}
}
\description{
Generate approximate pseudo-bulk data by random projections
}
