% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{asap_regression}
\alias{asap_regression}
\title{Poisson regression to estimate factor loading}
\usage{
asap_regression(
  Y_,
  log_x,
  a0 = 1,
  b0 = 1,
  max_iter = 10L,
  do_log1p = FALSE,
  verbose = FALSE,
  do_stdize_x = FALSE,
  std_topic_latent = FALSE
)
}
\arguments{
\item{log_x}{D x K log dictionary/design matrix}

\item{a0}{gamma(a0, b0)}

\item{b0}{gamma(a0, b0)}

\item{do_log1p}{do log(1+y) transformation}

\item{verbose}{verbosity (default: false)}

\item{std_topic_latent}{standardization of latent variables}

\item{Y}{D x N data matrix}

\item{do_stdize}{do the standardization of log_x}
}
\description{
Poisson regression to estimate factor loading
}
