% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{asap_regression}
\alias{asap_regression}
\title{Poisson regression to estimate factor loading}
\usage{
asap_regression(
  Y_,
  log_x,
  r_batch_effect = NULL,
  a0 = 1,
  b0 = 1,
  max_iter = 10L,
  do_scale = FALSE,
  do_log1p = FALSE,
  verbose = TRUE,
  do_stdize_x = FALSE
)
}
\arguments{
\item{log_x}{D x K log dictionary/design matrix}

\item{r_batch_effect}{D x B batch effect matrix (default: NULL)}

\item{a0}{gamma(a0, b0) (default: 1)}

\item{b0}{gamma(a0, b0) (default: 1)}

\item{do_scale}{scale each column by standard deviation (default: TRUE)}

\item{do_log1p}{do log(1+y) transformation (default: FALSE)}

\item{verbose}{verbosity (default: false)}

\item{Y}{D x N data matrix}

\item{do_stdize}{do the standardization of log_x}
}
\description{
Poisson regression to estimate factor loading
}
