% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{asap_regression_mtx}
\alias{asap_regression_mtx}
\title{Poisson regression to estimate factor loading}
\usage{
asap_regression_mtx(
  mtx_file,
  mtx_idx_file,
  log_x,
  r_batch_effect = NULL,
  r_x_row_names = NULL,
  r_mtx_row_names = NULL,
  r_taboo_names = NULL,
  a0 = 1,
  b0 = 1,
  max_iter = 10L,
  do_log1p = FALSE,
  verbose = FALSE,
  NUM_THREADS = 1L,
  BLOCK_SIZE = 100L,
  max_depth = 10000,
  do_stdize_x = FALSE
)
}
\arguments{
\item{mtx_file}{matrix-market-formatted data file (D x N, bgzip)}

\item{mtx_idx_file}{matrix-market colum index file}

\item{log_x}{D x K log dictionary/design matrix}

\item{r_batch_effect}{D x B batch effect matrix (default: NULL)}

\item{r_x_row_names}{(default: NULL)}

\item{r_mtx_row_names}{(default: NULL)}

\item{a0}{gamma(a0, b0)}

\item{b0}{gamma(a0, b0)}

\item{do_log1p}{do log(1+y) transformation}

\item{verbose}{verbosity}

\item{NUM_THREADS}{number of threads in data reading}

\item{BLOCK_SIZE}{disk I/O block size (number of columns)}

\item{max_depth}{maximum depth per column}

\item{do_stdize}{do the standardization of log_x}
}
\description{
Poisson regression to estimate factor loading
}
