% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Methods.R
\name{fit.topic.asap}
\alias{fit.topic.asap}
\title{Fit a topic model by ASAP}
\usage{
fit.topic.asap(
  mtx.file,
  k,
  pb.factors = 10,
  num.proj = 1,
  em.step = 100,
  max.pb.size = 1000,
  .burnin = 10,
  .reg.steps = 10,
  .reg.stdize = TRUE,
  .reg.stdize.latent = FALSE,
  a0 = 1,
  b0 = 1,
  index.file = paste0(mtx.file, ".index"),
  verbose = TRUE,
  num.threads = 1,
  block.size = 100,
  .rand.seed = 42,
  .eps = 1e-06,
  rate.m = 1,
  rate.v = 1,
  svd.init = TRUE,
  do.log1p = FALSE,
  max.depth = 10000,
  covar = NULL
)
}
\arguments{
\item{mtx.file}{data matrix file in a matrix market format}

\item{k}{number of topics}

\item{num.proj}{the number of pseudobulk projection steps (default: 1)}

\item{em.step}{Monte Carlo EM steps (default: 100)}

\item{max.pb.size}{maximum pseudobulk size (default: 1000)}

\item{.burnin}{burn-in period in the record keeping (default: 10)}

\item{.reg.steps}{number of steps in regression analysis (default: 10)}

\item{.reg.stdize}{standardize X in regression (default: TRUE)}

\item{.reg.stdize.latent}{standardize latent states in regression (default: FALSE)}

\item{a0}{Gamma(a0, b0) prior (default: 1)}

\item{b0}{Gamma(a0, b0) prior (default: 1)}

\item{index.file}{a file for column indexes (default: "{mtx.file}.index")}

\item{verbose}{verbosity}

\item{num.threads}{number of threads (default: 1)}

\item{block.size}{a block size for disk I/O (default: 100)}

\item{.rand.seed}{random seed (default: 42)}

\item{svd.init}{(default: TRUE)}

\item{do.log1p}{(default: FALSE)}

\item{max.depth}{(default: 1e4)}

\item{covar}{covariates (default: NULL)}

\item{eval.llik}{evaluate log-likelihood trace in PMF (default: TRUE)}
}
\value{
a list that contains:
\itemize{
\item \code{beta.rescaled} rescaled dictionary (feature x factor)
\item \code{theta.rescaled} rescaled loading (sample x factor)
\item \code{beta} dictionary (feature x factor)
\item \code{log.beta} log-dictionary (feature x factor)
\item \code{theta} loading (sample x factor)
\item \code{log.theta} log-loading (sample x factor)
\item \code{Y} pseudobulk (PB) data (feature x pseudobulk samples)
\item \code{nmf} non-negative matrix factorization results
\itemize{
\item \code{log.likelihood} log-likelihood trace
\item \code{beta} dictionary of PB data
\item \code{log.beta} log dictionary of PB data
\item \code{theta} factor loadings of PB data
\item \code{log.theta} log factor loadings of PB data
\item \code{log.phi} auxiliary variable (feature x factor)
\item \code{log.rho} auxiliary variable (sample x factor)
\item \code{beta.rescaled} rescaled dictionary
\item \code{theta.rescaled} rescaled factor loading
\item \code{depth}
}
\item \code{corr} sample x factor correlation matrix
}
}
\description{
Fit a topic model by ASAP
}
