% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{mmutil_read_rows_columns}
\alias{mmutil_read_rows_columns}
\title{Read a subset of rows and columns from the data matrix}
\usage{
mmutil_read_rows_columns(
  mtx_file,
  memory_location,
  r_row_index,
  r_column_index,
  verbose = FALSE
)
}
\arguments{
\item{mtx_file}{data file}

\item{memory_location}{column -> memory location}

\item{r_row_index}{row indexes to retrieve (1-based)}

\item{r_column_index}{column indexes to retrieve (1-based)}

\item{verbose}{verbosity}
}
\value{
a dense sub-matrix
}
\description{
Read a subset of rows and columns from the data matrix
}
\examples{

rr <- rgamma(100, 1, 1) # one hundred cells
mm <- matrix(rgamma(10 * 3, 1, 1), 10, 3)
data.hdr <- "test_sim"
.files <- asapR::mmutil_simulate_poisson(mm, rr, data.hdr)
data.file <- .files$mtx
idx.file <- .files$idx
mtx.idx <- asapR::mmutil_read_index(idx.file)
Y <- as.matrix(Matrix::readMM(data.file))
col.pos <- c(1,13,77) # 1-based
row.pos <- 1:10
yy <- asapR::mmutil_read_rows_columns(
                 data.file, mtx.idx, row.pos, col.pos)
all(Y[, col.pos, drop = FALSE] == yy)
print(head(Y[, col.pos, drop = FALSE]))
print(head(yy))
print(tail(yy))
unlink(list.files(pattern = data.hdr))

}
